Refactor: extract MemberInfo component
