+++ src/components/shared/MembersList/MemberInfo/MemberInfo.css
@@ -1,0 +1,63 @@
+@value query700 from '~styles/queries.css';
+
+.section {
+  padding: 20px 15px;
+}
+
+.section + .section {
+  padding-left: 0;
+}
+
+.usernameSection {
+  composes: section;
+  flex-grow: 0.5;
+  margin-right: auto;
+  width: 200px;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.displayName {
+  font-weight: var(--weight-bold);
+  color: var(--grey-1);
+}
+
+.displayName + .username {
+  display: inline-flex;
+  margin-left: 5px;
+}
+
+.address {
+  display: flex;
+  align-items: center;
+  line-height: 13px;
+}
+
+.address span {
+  font-size: var(--size-normal);
+  font-weight: 500;
+}
+
+.whitelistedIcon {
+  margin-left: 10px;
+  position: relative;
+}
+
+.whitelistedIcon svg {
+  fill: var(--primary);
+}
+
+.whitelistedIconTooltip {
+  width: 130px;
+}
+
+@media screen and query700 {
+  .section {
+    order: -1;
+  }
+
+  .usernameSection {
+    width: auto;
+  }
+}
