+++ src/components/shared/MembersList/MemberInfo/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MemberInfo';
