+++ src/components/shared/MembersList/MembersListItem.css.d.ts
@@ -1,7 +1,5 @@
 declare namespace MembersListItemCssNamespace {
   export interface IMembersListItemCss {
-    address: string;
-    displayName: string;
     main: string;
     mappings: string;
     names: string;
@@ -13,11 +11,7 @@ declare namespace MembersListItemCssNamespace {
     sourcesContent: string;
     stateHasReputation: string;
     stateReputationLoaded: string;
-    username: string;
-    usernameSection: string;
     version: string;
-    whitelistedIcon: string;
-    whitelistedIconTooltip: string;
   }
 }
